Reorganize routes
