+++ server/src/router/admin/thing/index.ts
